requirements.txt
bs4==0.0.1
aiohttp==2.2.5
async-timeout==2.0.0
chardet==3.0.4
feedparser==5.2.1
multidict==3.2.0
python-telegram-bot==8.1.1
python2==1.2
telepot==12.3
yarl==0.13.0
cached-property==1.3.1
docker==2.5.1
docker-compose==1.16.1
docker-pycreds==0.2.1
dockerpty==0.4.1
docopt==0.6.2
jsonschema==2.6.0
PyYAML==3.12
requests==2.11.1
six==1.11.0
texttable==0.9.1
websocket-client==0.44.0
